translation-service/requirements.txt
# Core dependencies (shared from §14 of Shared_Contracts.md)
python==3.11.*
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*
httpx==0.28.*
structlog==24.4.*
prometheus-client==0.21.*
redis==5.2.*
python-multipart==0.0.18

# AI4Bharat IndicTrans2 and dependencies
transformers==4.45.2
torch==2.1.2
sentencepiece==0.1.99
sacremoses==0.0.53

# Language detection
langid==1.1.6
fasttext==0.9.2

# Transliteration and script handling
indic-nlp-library==0.90

# Utility
numpy==1.24.3
requests==2.31.0
